# FE/aiPhaseII/src/api/ai二期.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ai二期",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "用户登录",
        "deprecated": false,
        "description": "用户登录并返回访问令牌",
        "operationId": "login_api_auth_login_post",
        "tags": [
          "认证服务"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "\n        返回数据结构:\n        ```json\n        {\n            \"code\": 200,\n            \"message\": \"登录成功\",\n            \"data\": {\n                \"access_token\": \"访问令牌\",\n                \"token_type\": \"令牌类型\",\n                \"expires_in\": \"过期时间(秒)\"\n            }\n        }\n        ```\n        ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "退出登录",
        "deprecated": false,
        "description": "使当前用户登录的令牌失效",
        "operationId": "logout_api_auth_logout_post",
        "tags": [
          "认证服务"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": " ",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LogoutRequest"
                  }
                ],
                "default": {}
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "\n        返回数据结构:\n        ```json\n        {\n            \"code\": 200,\n            \"message\": \"退出登录成功\",\n            \"data\": null\n        }\n        ```\n        ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/auth/profile": {
      "get": {
        "summary": "获取个人信息",
        "deprecated": false,
        "description": "获取当前登录用户的详细信息",
        "operationId": "get_profile_api_auth_profile_get",
        "tags": [
          "认证服务"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "\n        返回数据结构:\n        ```json\n        {\n            \"code\": 200,\n            \"message\": \"获取成功\",\n            \"data\": {\n                \"id\": \"用户ID\",\n                \"username\": \"用户名\",\n                \"realName\": \"真实姓名\",\n                \"phone\": \"手机号\",\n                \"email\": \"邮箱\",\n                \"enterpriseId\": \"企业ID\",\n                \"enterpriseName\": \"企业名称\",\n                \"enterpriseRegionCode\": \"企业所在地区编码\",\n                \"enterpriseRegionName\": \"企业所在地区名称\",\n                \"enterpriseAddress\": \"企业详细地址\",\n                \"role\": \"角色\",\n                \"status\": \"状态(0:待审核,1:正常,2:禁用)\",\n                \"avatar\": \"头像URL\",\n                \"createdAt\": \"创建时间\",\n                \"updatedAt\": \"更新时间\"\n            }\n        }\n        ```\n        ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "更新个人信息",
        "deprecated": false,
        "description": "更新当前登录用户的个人信息（姓名、邮箱、头像），手机号作为登录账号不可修改",
        "operationId": "update_user_info_api_auth_profile_put",
        "tags": [
          "认证服务"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserInfoRequest"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "\n        返回数据结构:\n        ```json\n        {\n            \"code\": 200,\n            \"message\": \"更新成功\",\n            \"data\": {\n                \"id\": \"用户ID\",\n                \"name\": \"用户姓名\",\n                \"account\": \"账号\",\n                \"phone\": \"手机号\",\n                \"email\": \"邮箱\",\n                \"avatar\": \"头像URL\",\n                \"enterpriseId\": \"企业ID\",\n                \"enterpriseName\": \"企业名称\",\n                \"enterpriseRegionCode\": \"企业所在地区编码\",\n                \"enterpriseRegionName\": \"企业所在地区名称\",\n                \"enterpriseAddress\": \"企业详细地址\",\n                \"departmentId\": \"部门ID\",\n                \"departmentName\": \"部门名称\",\n                \"userType\": \"用户类型\",\n                \"status\": \"状态\"\n            }\n        }\n        ```\n        ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/auth/upload": {
      "post": {
        "summary": "上传文件",
        "deprecated": false,
        "description": "上传文件并返回访问URL",
        "operationId": "upload_file_api_auth_upload_post",
        "tags": [
          "认证服务"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "\n        返回数据结构:\n        ```json\n        {\n            \"code\": 200,\n            \"message\": \"上传成功\",\n            \"data\": {\n                \"url\": \"文件访问URL\",\n                \"filename\": \"原始文件名\",\n                \"content_type\": \"文件类型\"\n            }\n        }\n        ```\n        ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/auth/changePassword": {
      "post": {
        "summary": "修改密码",
        "deprecated": false,
        "description": "使用原密码修改为新密码",
        "operationId": "change_password_api_auth_changePassword_post",
        "tags": [
          "认证服务"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "\n        返回数据结构:\n        ```json\n        {\n            \"code\": 200,\n            \"message\": \"密码修改成功\",\n            \"data\": null\n        }\n        ```\n        ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/auth/resetPassword": {
      "post": {
        "summary": "重置密码",
        "deprecated": false,
        "description": "重置用户密码为123456",
        "operationId": "reset_password_api_auth_resetPassword_post",
        "tags": [
          "认证服务"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSchema"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "title": "ChangePasswordRequest",
        "required": [
          "oldPassword",
          "newPassword",
          "newPasswordConfirm"
        ],
        "type": "object",
        "properties": {
          "oldPassword": {
            "title": "Oldpassword",
            "type": "string",
            "description": "原密码"
          },
          "newPassword": {
            "title": "Newpassword",
            "maxLength": 16,
            "minLength": 6,
            "type": "string",
            "description": "新密码"
          },
          "newPasswordConfirm": {
            "title": "Newpasswordconfirm",
            "type": "string",
            "description": "确认新密码"
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "LoginRequest": {
        "title": "LoginRequest",
        "required": [
          "username",
          "password"
        ],
        "type": "object",
        "properties": {
          "username": {
            "title": "Username",
            "type": "string",
            "description": "账号"
          },
          "password": {
            "title": "Password",
            "type": "string",
            "description": "密码"
          }
        }
      },
...
      "LogoutRequest": {
        "title": "LogoutRequest",
        "type": "object",
        "properties": {}
      },
...
      "ResetPasswordRequest": {
        "title": "ResetPasswordRequest",
        "type": "object",
        "properties": {
          "userId": {
            "title": "User Id",
            "type": "integer",
            "description": "用户ID"
          }
        }
      },
      "ResponseSchema": {
        "title": "ResponseSchema",
        "type": "object",
        "properties": {
          "code": {
            "title": "Code",
            "type": "integer",
            "default": 200
          },
          "data": {
            "title": "Data"
          },
          "message": {
            "title": "Message",
            "type": "string",
            "default": "请求成功"
          }
        }
      },
...
      "UpdateUserInfoRequest": {
        "title": "UpdateUserInfoRequest",
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "description": "用户姓名"
          },
          "email": {
            "title": "Email",
            "type": "string",
            "description": "邮箱",
            "format": "email"
          },
          "avatar": {
            "title": "Avatar",
            "type": "string",
            "description": "头像URL"
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}